moved the code that translates preview templates from notification-utils up to the get_template function so it can be used everywheree
