+++ app/main/views/templates.py
@@ -70,23 +70,7 @@ def get_email_preview_template(template, template_id, service_id):
         show_recipient=True,
         page_count=get_page_count_for_letter(template),
     )
-    template_str = str(email_preview_template)
-    translate = {
-        "From": _("From"),
-        "To": _("To"),
-        "Subject": _("Subject")
-    }
 
-    def translate_brackets(x):
-        g = x.group(0)
-        english = g[1:-1]  # drop brackets
-        if english not in translate:
-            return english
-        return translate[english]
-
-    # this regex finds test inside []
-    template_str = re.sub(r"\[[^]]*\]", translate_brackets, template_str)
-    email_preview_template.html = template_str
     return email_preview_template
