+++ app/templates/views/check/column-errors.html
@@ -223,6 +223,6 @@
   {% endif %}
 
   <h2 class="heading-medium">{{ _('Preview of') }} {{ template.name }}</h2>
-  {{ template|string }}
+  {{ template.html|safe }}
 
 {% endblock %}
