+++ app/templates/views/templates/_template_history.html
@@ -12,5 +12,6 @@
 </div>
 
 <div class="column-whole">
-  {{ template|string }}
+  {{ template.html|safe }}
+
 </div>
