+++ app/templates/views/templates/start-tour.html
@@ -12,7 +12,7 @@
 
   <h1 class="heading-large">{{ template.name }}</h1>
 
-  {{ template|string }}
+  {{ template.html|safe }}
 
   <div class="page-footer">
     <a href="{{ url_for('.send_test', service_id=current_service.id, template_id=template.id, help=2) }}" class="button">Continue</a>
